requirements.txt
Pillow==10.4.0
pywin32==307
requests==2.32.3
urllib3==2.2.2
websockets==13.0.1
WMI==1.5.1